@@ -1,72 +1,82 @@
 {
   "description": "Maison Saine",
-  "license": "SEE LICENSE IN <your-license-filename>",
+  "license": "MIT",
   "readme": "NativeScript Application",
-  "repository": "<fill-your-repository-here>",
+  "repository": "https://github.com/Akylas/mainson_saine",
   "nativescript": {
     "id": "com.akylas.maisonsaine",
     "tns-android": {
-      "version": "5.1.0"
+      "version": "6.4.1"
     },
     "tns-ios": {
-      "version": "5.1.0"
+      "version": "6.4.2"
     }
   },
   "scripts": {
     "pandroid": "cross-var tns build android --bundle --release --keyStorePath $KEYSTORE_PATH --keyStorePassword $KEYSTORE_PASS --keyStoreAlias $KEYSTORE_ALIAS --keyStoreAliasPassword $KEYSTORE_PASS --env.uglify --env.snapshot --env.production",
     "randroid": "cross-var tns run android --bundle --release --keyStorePath $KEYSTORE_PATH --keyStorePassword $KEYSTORE_PASS --keyStoreAlias $KEYSTORE_ALIAS --keyStoreAliasPassword $KEYSTORE_PASS --env.snapshot --env.production",
-    "pios": "tns build ios --bundle --release --env.uglify --env.production --for-device --provision 'Maison Saine'"
+    "pios": "tns build ios --bundle --release --env.uglify --env.production --for-device --provision 'Maison Saine'",
+    "ios.generate_licences": "node generate_ios_licences.js",
+    "android.generate_licences": "cd platforms/android; ./gradlew  generateLicenseReport"
   },
   "dependencies": {
-    "@tweenjs/tween.js": "17.2.0",
-    "nativescript-collectionview": "1.0.17",
+    "@mdi/font": "4.9.95",
+    "@nativescript/core": "6.5.1",
+    "nativescript-collectionview": "2.0.21",
+    "nativescript-extendedinfo": "^1.0.7",
     "nativescript-fonticon": "~2.0.0",
-    "nativescript-htmllabel": "1.0.4",
-    "nativescript-image": "^2.0.1",
-    "nativescript-material-button": "2.0.2",
-    "nativescript-material-cardview": "2.0.2",
-    "nativescript-material-ripple": "2.0.2",
-    "nativescript-photoviewer": "^1.4.0",
-    "nativescript-ui-listview": "^5.0.1",
-    "nativescript-vue": "^2.0.2",
-    "tinycolor2": "^1.4.1",
-    "tns-core-modules": "5.1.1"
+    "nativescript-htmllabel": "1.0.32",
+    "nativescript-image": "^3.0.6",
+    "nativescript-inappbrowser": "^2.3.0",
+    "nativescript-material-bottomsheet": "^3.2.3",
+    "nativescript-material-button": "3.2.3",
+    "nativescript-material-cardview": "3.2.3",
+    "nativescript-material-ripple": "3.2.3",
+    "nativescript-photoviewer": "^2.1.5",
+    "nativescript-tween": "0.0.6",
+    "nativescript-vue": "^2.6.1",
+    "tinycolor2": "^1.4.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.2.2",
-    "@babel/preset-env": "^7.2.3",
+    "@babel/core": "^7.9.6",
+    "@babel/preset-env": "^7.9.6",
+    "@sentry/webpack-plugin": "^1.11.1",
     "@types/tinycolor": "^1.1.27",
-    "@types/tween.js": "^16.9.0",
-    "babel-loader": "^8.0.4",
+    "@types/tween.js": "^18.5.0",
+    "babel-loader": "^8.1.0",
     "babel-traverse": "6.26.0",
     "babel-types": "6.26.0",
     "babylon": "6.18.0",
-    "clean-webpack-plugin": "^1.0.0",
-    "copy-webpack-plugin": "^4.6.0",
+    "clean-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^5.1.1",
     "cross-var": "^1.1.0",
-    "css-hot-loader": "^1.4.3",
-    "css-loader": "^2.1.0",
+    "css-hot-loader": "^1.4.4",
+    "css-loader": "^3.5.3",
     "lazy": "1.0.11",
-    "mini-css-extract-plugin": "^0.5.0",
-    "nativescript-dev-webpack": "^0.18.5",
-    "nativescript-vue-template-compiler": "^2.0.2",
-    "nativescript-worker-loader": "~0.9.2",
-    "node-sass": "^4.11.0",
-    "sass-loader": "^7.1.0",
-    "terser-webpack-plugin": "^1.2.1",
-    "tns-platform-declarations": "^5.1.1",
-    "ts-loader": "^5.3.2",
-    "tslint": "^5.12.0",
-    "tslint-config-prettier": "^1.17.0",
-    "tslint-config-standard": "^8.0.1",
-    "typescript": "^3.2.2",
-    "uglifyjs-webpack-plugin": "~2.1.1",
-    "vue": "^2.5.21",
-    "vue-loader": "^15.4.2",
-    "vue-property-decorator": "^7.2.0",
-    "webpack": "^4.28.2",
-    "webpack-bundle-analyzer": "~3.0.3",
-    "webpack-cli": "^3.1.2"
+    "merge-options": "^2.0.0",
+    "mini-css-extract-plugin": "^0.9.0",
+    "nativescript-dev-webpack": "^1.5.1",
+    "nativescript-vue-template-compiler": "^2.6.1",
+    "nativescript-worker-loader": "~0.11.0",
+    "node-sass": "^4.14.1",
+    "plist": "^3.0.1",
+    "sass-loader": "^8.0.2",
+    "scss-symbols-parser": "^2.0.1",
+    "string-replace-loader": "^2.3.0",
+    "terser-webpack-plugin": "^3.0.1",
+    "tns-platform-declarations": "^6.5.1",
+    "ts-loader": "^7.0.4",
+    "tslint": "^6.1.2",
+    "tslint-config-prettier": "^1.18.0",
+    "tslint-config-standard": "^9.0.0",
+    "typescript": "^3.9.2",
+    "uglifyjs-webpack-plugin": "~2.2.0",
+    "vue": "^2.6.11",
+    "vue-loader": "^15.9.2",
+    "vue-property-decorator": "^8.4.2",
+    "webpack": "^4.43.0",
+    "webpack-bundle-analyzer": "~3.7.0",
+    "webpack-cli": "^3.3.11"
   },
   "author": "Martin Guillon"
 }
