@@ -1,7 +1,7 @@
 import Vue from 'nativescript-vue';
 import { Prop } from 'vue-property-decorator';
-import { Color, Page } from 'tns-core-modules/ui/page/page';
-import { isAndroid, isIOS } from 'tns-core-modules/platform/platform';
+import { Color, Page } from '@nativescript/core/ui/page/page';
+import { isAndroid, isIOS } from '@nativescript/core/platform/platform';
 
 import { actionBarHeight, darkColor, primaryColor } from '../variables';
 
@@ -16,19 +16,22 @@ export default class BaseVueComponent extends Vue {
     darkColor;
     public actionBarHeight = actionBarHeight;
     get page() {
-        return (this.$refs.page as any).nativeView as Page;
+        return this.$refs.page && (this.$refs.page as any).nativeView as Page;
     }
     mounted() {
         const page = this.page;
-        // page.backgroundSpanUnderStatusBar = true;
-        page.actionBarHidden = true;
-        if (isIOS) {
-            page.statusBarStyle = 'dark';
-        } else {
-            page.androidStatusBarBackground = null;
-            page.androidStatusBarBackground = new Color(this.darkColor);
+        if (page) {
+            // page.backgroundSpanUnderStatusBar = true;
+            page.actionBarHidden = true;
+            if (isIOS) {
+                page.statusBarStyle = 'dark';
+            } else {
+                page.androidStatusBarBackground = null;
+                page.androidStatusBarBackground = new Color(this.darkColor);
+            }
+            page.backgroundColor = this.themeColor;
+
         }
-        page.backgroundColor = this.themeColor;
     }
 
     // get darkColor() {
