@@ -2,22 +2,21 @@
     <Page ref="page" class="page themedBack" @navigatedTo="onNavigatedTo">
         <StackLayout backgroundColor="#F4F4F4">
             <StackLayout orientation="horizontal" class="actionBar">
-                <MDCButton class="actionBarButton" variant="text" :text="'mdi-arrow-left' | fonticon" @tap="onTap('back', $event)" />
+                <Button class="actionBarButton" variant="text" text="mdi-arrow-left" @tap="onTap('back', $event)" />
                 <Label class="actionBarTitle" verticalAlignment="center" :text="'lexique' | uppercase" color="white" />
             </StackLayout>
-            <RadListView width="100%" height="100%" :items="dataItems" :itemTemplateSelector="templateSelector">
+            <CollectionView width="100%" height="100%" :items="dataItems" :itemTemplateSelector="templateSelector" :itemIdGenerator="(item,i)=>i">
                 <v-template name="section">
                     <GridLayout rows="10,auto" columns="15,*" backgroundColor="white">
                         <Label row="1" col="1" :text="item.name" :color="darkColor" fontSize="18" class="nunitobold" borderBottomWidth="2" :borderBottomColor="darkColor" />
                     </GridLayout>
                 </v-template>
                 <v-template name="item">
-                    <GridLayout backgroundColor="white">
-                        <Label padding="10" verticalAlignment="center" :text="item.name" fontSize="18" class="nunito" color="#666"></Label>
-                        <Ripple :rippleColor="themeColor" @tap="onNavigationItemTap(item)" />
+                    <GridLayout backgroundColor="white" :rippleColor="themeColor" @tap="onNavigationItemTap(item)" >
+                        <Label padding="10" verticalAlignment="center" :text="item.name" fontSize="18" color="#666"></Label>
                     </GridLayout>
                 </v-template>
-            </RadListView>
+            </CollectionView>
         </StackLayout>
     </Page>
 </template>
@@ -25,11 +24,11 @@
 <script lang="ts">
 import BaseVueComponent from './BaseVueComponent';
 import { Component, Prop } from 'vue-property-decorator';
-import { screen } from 'platform';
-import { ObservableArray } from 'data/observable-array/observable-array';
-import { EventData, NavigatedData, Observable } from 'tns-core-modules/ui/page/page';
-import * as http from 'tns-core-modules/http';
-import * as fileSystem from 'tns-core-modules/file-system';
+import { screen } from '@nativescript/core/platform';
+import { ObservableArray } from '@nativescript/core/data/observable-array/observable-array';
+import { EventData, NavigatedData, Observable } from '@nativescript/core/ui/page/page';
+import * as http from '@nativescript/core/http';
+import * as fileSystem from '@nativescript/core/file-system';
 
 interface Data {}
 
