@@ -1,15 +1,15 @@
 <template>
     <Page ref="page" class="page">
         <GridLayout rows="auto,*" backgroundColor="white" iosOverflowSafeArea="false">
-            <StackLayout row="0" orientation="horizontal" :backgroundColor="darkColor">
-                <MDCButton class="actionBarButton" variant="text" :text="'mdi-arrow-left' | fonticon" @tap="onTap('back', $event)" />
+            <StackLayout row="0" orientation="horizontal" class="actionBar">
+                <Button class="actionBarButton" variant="text" text="mdi-arrow-left" @tap="onTap('back', $event)" />
                 <Label class="actionBarTitle" verticalAlignment="center" :text="lexiqueData['Agent pathogène']" color="white" />
             </StackLayout>
             <ScrollView  row="1" >
                 <StackLayout>
                 <StackLayout padding="15 15 0 15" v-for="(key) in Object.keys(lexiqueData).filter(k=>k !== 'Agent pathogène')" :key="key">
                      <Label :text="key" :color="darkColor" fontSize="18" class="nunitobold" borderBottomWidth="2" :borderBottomColor="darkColor"/>
-                    <HTMLLabel fontSize="15" class="nunito" padding="5" :html="Array.isArray(lexiqueData[key])?lexiqueData[key].join(', '):lexiqueData[key]" />
+                    <Label fontSize="15" padding="5" :html="Array.isArray(lexiqueData[key])?lexiqueData[key].join(', '):lexiqueData[key]" />
                 </StackLayout>
                 </StackLayout>
             </ScrollView>
@@ -20,10 +20,10 @@
 <script lang="ts">
 import BaseVueComponent from './BaseVueComponent';
 import { Component, Prop } from 'vue-property-decorator';
-import { screen } from 'platform';
-import { ObservableArray } from 'data/observable-array/observable-array';
-import { Image } from 'tns-core-modules/ui/image/image';
-import { EventData } from 'tns-core-modules/ui/page/page';
+import { screen } from '@nativescript/core/platform';
+import { ObservableArray } from '@nativescript/core/data/observable-array/observable-array';
+import { Image } from '@nativescript/core/ui/image/image';
+import { EventData } from '@nativescript/core/ui/page/page';
 
 import { actionBarHeight, backgroundColor, darkColor, roomHeaderHeight } from '../variables';
 
