@@ -2,16 +2,16 @@
     <Page ref="page" class="page themedBack">
         <StackLayout>
             <StackLayout orientation="horizontal" class="actionBar">
-                <MDCButton class="actionBarButton" variant="text" :text="'mdi-arrow-left' | fonticon" @tap="onTap('back', $event)" />
+                <Button class="actionBarButton" variant="text" text="mdi-arrow-left" @tap="onTap('back', $event)" />
                 <Label class="actionBarTitle" verticalAlignment="center" :text="'recettes' | uppercase" color="white" />
             </StackLayout>
-            <CollectionView width="100%" height="100%" :items="dataItems" backgroundColor="#F4F4F4" :rowHeight="itemHeight" colWidth="49.8%">
+            <CollectionView width="100%" height="100%" :items="dataItems" backgroundColor="#F4F4F4" :rowHeight="itemHeight" colWidth="49.8%" :itemIdGenerator="(item,i)=>i">
                 <v-template>
                     <GridLayout width="100%" height="100%" rows="*" columns="*">
                         <CardView margin="10" borderRadius="4" @tap="onNavigationItemTap(item)" :rippleColor="themeColor">
                             <StackLayout width="100%" height="100%" orientation="vertical" isUserInteractionEnabled="false">
-                                <AImage row="0" borderRadius="4 4 0 0" :height="itemWidth" :src="item.image" stretch="aspectFill" />
-                                <HTMLLabel row="1" class="nunito" fontSize="15" width="100%" padding="5" color="black" :html="item.title" isUserInteractionEnabled="false" backgroundColor="transparent" />
+                                <Image row="0" borderRadius="4 4 0 0" :height="itemWidth" :src="item.image" stretch="aspectFill" fadeDuration="200"/>
+                                <Label row="1" fontSize="15" width="100%" padding="5" color="black" :html="item.title" isUserInteractionEnabled="false" backgroundColor="transparent" />
                             </StackLayout>
                         </CardView>
                     </GridLayout>
@@ -24,10 +24,10 @@
 <script lang="ts">
 import BaseVueComponent from './BaseVueComponent';
 import { Component, Prop } from 'vue-property-decorator';
-import { screen } from 'platform';
-import { ObservableArray } from 'data/observable-array/observable-array';
+import { screen } from '@nativescript/core/platform';
+import { ObservableArray } from '@nativescript/core/data/observable-array/observable-array';
 import { RecetteData } from '~/data/data';
-import { EventData } from 'tns-core-modules/ui/page/page';
+import { EventData } from '@nativescript/core/ui/page/page';
 import { getRecettes } from '~/services/data.item.service';
 
 @Component({})
