@@ -1,8 +1,8 @@
 // const dev = TNS_ENV === 'development';
 const dev = false;
-import Vue, { registerElement } from 'nativescript-vue';
+import Vue from 'nativescript-vue';
 
-import * as trace from 'tns-core-modules/trace';
+import * as trace from '@nativescript/core/trace';
 const errorHandler: trace.ErrorHandler = {
     handlerError(err) {
         // (development 1)
@@ -20,7 +20,7 @@ trace.setErrorHandler(errorHandler);
 
 // naming main file app is important for snapshot
 import './app.scss';
-import { isAndroid, isIOS } from 'platform';
+import { isAndroid, isIOS } from '@nativescript/core/platform';
 // import { addCategories, categories, enable } from 'trace';
 // addCategories(categories.NativeLifecycle);
 // addCategories(categories.ViewHierarchy);
@@ -30,36 +30,36 @@ import { isAndroid, isIOS } from 'platform';
 Vue.config.silent = !dev;
 Vue.config['debug'] = dev;
 
-import { fonticon, TNSFontIcon } from 'nativescript-fonticon';
-
 import CollectionView from 'nativescript-collectionview/vue';
 Vue.use(CollectionView);
-import RadListViewPlugin from 'nativescript-ui-listview/vue';
-Vue.use(RadListViewPlugin);
-
-registerElement('MDCButton', () => require('nativescript-material-button').Button);
+import {installMixins} from 'nativescript-material-core';
+installMixins();
+import { install as installBottomSheets } from 'nativescript-material-bottomsheet';
+installBottomSheets();
+Vue.registerElement('Button', () => require('nativescript-material-button').Button);
 // registerElement('FloatingActionButton', () => require('nativescript-material-components/floatingactionbutton').FloatingActionButton);
 // registerElement('MDCActivityIndicator', () => require('nativescript-material-components/activityindicator').ActivityIndicator);
-registerElement('CardView', () => require('nativescript-material-cardview').CardView);
-registerElement('Ripple', () => require('nativescript-material-ripple').Ripple);
-registerElement('HTMLLabel', () => require('nativescript-htmllabel').Label);
+Vue.registerElement('CardView', () => require('nativescript-material-cardview').CardView);
+Vue.registerElement('Ripple', () => require('nativescript-material-ripple').Ripple);
+import BottomSheetPlugin from 'nativescript-material-bottomsheet/vue';
+Vue.use(BottomSheetPlugin);
 
-import * as app from 'tns-core-modules/application';
+// import { addCategories, enable } from '@nativescript/core/trace';
+// import {CollectionViewTraceCategory} from 'nativescript-collectionview';
+// addCategories(CollectionViewTraceCategory);
+// enable();
+
+import { Label, enableIOSDTCoreText } from 'nativescript-htmllabel';
+enableIOSDTCoreText();
+Vue.registerElement('Label', () => Label);
+
 import * as imageModule from 'nativescript-image';
-registerElement('AImage', () => imageModule.Img);
-
-app.on(app.launchEvent, () => imageModule.initialize({ isDownsampleEnabled: true }), imageModule.getImagePipeline().clearCaches());
-app.on(app.exitEvent, args => imageModule.shutDown());
+Vue.registerElement('Image', () => imageModule.Img);
+imageModule.initialize({ isDownsampleEnabled: true });
+// app.on(app.launchEvent, () => imageModule.initialize({ isDownsampleEnabled: true }), imageModule.getImagePipeline().clearCaches());
+// app.on(app.exitEvent, args => imageModule.shutDown());
 
 import Home from './components/Home.vue';
-
-// Prints all icon classes loaded
-// TNSFontIcon.debug = true;
-TNSFontIcon.paths = {
-    mdi: './assets/mdi.css'
-};
-TNSFontIcon.loadCss();
-Vue.filter('fonticon', fonticon);
 
 Vue.filter('uppercase', function(value) {
     if (!value) return '';
