@@ -1,5 +1,4 @@
-import Vue from "vue"
-
+import Vue from 'vue';
 declare module "vue/types/vue" {
     // 3. Declare augmentation for Vue
     interface Vue {
