@@ -1,134 +1,245 @@
-const { relative, resolve, sep } = require("path")
+const { join, relative, resolve, sep } = require('path');
+const { existsSync, mkdirSync, readFileSync } = require('fs');
 
-const webpack = require("webpack")
-const CleanWebpackPlugin = require("clean-webpack-plugin")
-const CopyWebpackPlugin = require("copy-webpack-plugin")
-const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer")
-const TerserPlugin = require("terser-webpack-plugin")
+const webpack = require('webpack');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
+const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
+const TerserPlugin = require('terser-webpack-plugin');
 
-const VueLoaderPlugin = require("vue-loader/lib/plugin")
+const { VueLoaderPlugin } = require('vue-loader');
 
-const NsVueTemplateCompiler = require("nativescript-vue-template-compiler")
-// NsVueTemplateCompiler.registerElement(
-//   'FormattedString',
-//   () => require('tns-core-modules/text/formatted-string').FormattedString
-// )
-// NsVueTemplateCompiler.registerElement('Span', () => require('tns-core-modules/text/span').Span)
-// NsVueTemplateCompiler.registerElement(
-//     "MDCTextField",
-//     () => require("~/nativescript-material-components/textfield").TextField,
-//     {
-//         model: {
-//             prop: "text",
-//             event: "textChange"
-//         }
-//     }
-// )
-// NsVueTemplateCompiler.registerElement(
-//     "TabView",
-//     () => require("~/tns-core-modules/ui/tab-view").TabView,
-//     {
-//         model: {
-//             prop: "selectedIndex",
-//             event: "selectedIndexChange"
-//         }
-//     }
-// )
+const nsWebpack = require('nativescript-dev-webpack');
+const nativescriptTarget = require('nativescript-dev-webpack/nativescript-target');
+const { NativeScriptWorkerPlugin } = require('nativescript-worker-loader/NativeScriptWorkerPlugin');
+const hashSalt = Date.now().toString();
+const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
+const mergeOptions = require('merge-options');
+const SentryCliPlugin = require('@sentry/webpack-plugin');
 
-const nsWebpack = require("nativescript-dev-webpack")
-const nativescriptTarget = require("nativescript-dev-webpack/nativescript-target")
-const {
-    NativeScriptWorkerPlugin
-} = require("nativescript-worker-loader/NativeScriptWorkerPlugin")
+// returns a new object with the values at each key mapped using mapFn(value)
 
-module.exports = env => {
-    // Add your custom Activities, Services and other android app components here.
-    const appComponents = [
-        "tns-core-modules/ui/frame",
-        "tns-core-modules/ui/frame/activity"
-    ]
+// temporary hack to support v-model using ns-vue-template-compiler
+// See https://github.com/nativescript-vue/nativescript-vue/issues/371
+const NsVueTemplateCompiler = require('nativescript-vue-template-compiler');
+NsVueTemplateCompiler.registerElement('MDTextField', () => require('nativescript-material-textfield').TextField, {
+    model: {
+        prop: 'text',
+        event: 'textChange'
+    }
+});
+NsVueTemplateCompiler.registerElement('MDSlider', () => require('nativescript-material-slider').Slider, {
+    model: {
+        prop: 'value',
+        event: 'valueChange'
+    }
+});
 
-    const platform = env && ((env.android && "android") || (env.ios && "ios"))
+module.exports = (env, params = {}) => {
+    const platform = env && ((env.android && 'android') || (env.ios && 'ios') || env.platform);
     if (!platform) {
-        throw new Error("You need to provide a target platform!")
+        throw new Error('You need to provide a target platform!');
     }
 
-    const platforms = ["ios", "android"]
-    const projectRoot = __dirname
+    // Add your custom Activities, Services and other android app components here.
+    let appComponents = env.appComponents || [];
+    appComponents.push(...['tns-core-modules/ui/frame', 'tns-core-modules/ui/frame/activity']);
 
-    const tsconfig = platform === "android" ? 'tsconfig.android.json' : 'tsconfig.json'
+    console.log('appComponents', appComponents);
+    const platforms = ['ios', 'android'];
+    const projectRoot = __dirname;
+
+    let tsconfig = params.tsconfig;
+    if (!tsconfig) {
+        tsconfig = 'tsconfig.json';
+        if (existsSync(resolve(projectRoot, `tsconfig.${platform}.json`))) {
+            tsconfig = `tsconfig.${platform}.json`;
+        }
+    }
+    console.log('tsconfig', tsconfig);
+
+    if (env.platform) {
+        platforms.push(env.platform);
+    }
 
     // Default destination inside platforms/<platform>/...
-    const dist = resolve(
-        projectRoot,
-        nsWebpack.getAppPath(platform, projectRoot)
-    )
-    const appResourcesPlatformDir = platform === "android" ? "Android" : "iOS"
+    const dist = resolve(projectRoot, nsWebpack.getAppPath(platform, projectRoot));
 
+    if (env.adhoc) {
+        env = Object.assign({}, env, {
+            production: true,
+            sentry: true,
+            sourceMap: true,
+            uglify: true
+        });
+    }
     const {
         // The 'appPath' and 'appResourcesPath' values are fetched from
-        // the nsconfig.json configuration file
-        // when bundling with `tns run android|ios --bundle`.
-        appPath = "app",
-        appResourcesPath = "app/App_Resources",
+        // the nsconfig.json configuration file.
+        appPath = '../common/app',
+        appResourcesPath = 'App_Resources',
 
         // You can provide the following flags when running 'tns run android|ios'
         snapshot, // --env.snapshot
-        uglify, // --env.uglify
         production, // --env.production
         report, // --env.report
+        hmr, // --env.hmr
         sourceMap, // --env.sourceMap
-        hmr // --env.hmr
-    } = env
+        hiddenSourceMap, // --env.hiddenSourceMap
+        inlineSourceMap, // --env.inlineSourceMap
+        unitTesting, // --env.unitTesting
+        verbose, // --env.verbose
+        snapshotInDocker, // --env.snapshotInDocker
+        skipSnapshotTools, // --env.skipSnapshotTools
+        compileSnapshot, // --env.compileSnapshot
+        uglify, // --env.uglify
+        devlog, // --env.devlog
+        sentry, // --env.sentry
+        adhoc // --env.adhoc
+    } = env;
 
-    const mode = production ? "production" : "development"
+    const useLibs = compileSnapshot;
+    const isAnySourceMapEnabled = !!sourceMap || !!hiddenSourceMap || !!inlineSourceMap;
+    const externals = nsWebpack.getConvertedExternals(env.externals);
 
-    const appFullPath = resolve(projectRoot, appPath)
-    const appResourcesFullPath = resolve(projectRoot, appResourcesPath)
+    const mode = production ? 'production' : 'development';
 
-    const entryModule = nsWebpack.getEntryModule(appFullPath)
-    const entryPath = `.${sep}${entryModule}`
+    const appFullPath = resolve(projectRoot, appPath);
+    const hasRootLevelScopedModules = nsWebpack.hasRootLevelScopedModules({ projectDir: projectRoot });
+    let coreModulesPackageName = 'tns-core-modules';
+    const alias = env.alias || {};
+    alias['~'] = appFullPath;
+    alias['@'] = appFullPath;
+    alias['vue'] = 'nativescript-vue';
+    Object.assign(alias, {
+        '../driver/oracle/OracleDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './oracle/OracleDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/cockroachdb/CockroachDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './cockroachdb/CockroachDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './cordova/CordovaDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './react-native/ReactNativeDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/react-native/ReactNativeDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './nativescript/NativescriptDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/nativescript/NativescriptDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './mysql/MysqlDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/mysql/MysqlDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './postgres/PostgresDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/postgres/PostgresDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './expo/ExpoDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './aurora-data-api/AuroraDataApiDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/aurora-data-api/AuroraDataApiDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './sqlite/SqliteDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/sqljs/SqljsDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './sqljs/SqljsDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/sqlserver/SqlServerDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './sqlserver/SqlServerDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './mongodb/MongoDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/mongodb/MongoDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        './cordova/CordovaDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver',
+        '../driver/cordova/CordovaDriver': 'nativescript-akylas-sqlite/typeorm/NativescriptDriver'
+    });
 
-    const config = {
-        mode: mode,
-        context: appFullPath,
-        // stats:'verbose',
-        watch: !!hmr,
-        watchOptions: !!hmr
-            ? {
-                  ignored: [
-                      appResourcesFullPath,
-                      // Don't watch hidden files
-                      "**/.*"
-                  ]
-              }
-            : undefined,
-        target: nativescriptTarget,
-        entry: {
-            bundle: entryPath
+    if (hasRootLevelScopedModules) {
+        coreModulesPackageName = '@nativescript/core';
+        alias['tns-core-modules'] = coreModulesPackageName;
+    }
+
+    const appResourcesFullPath = resolve(projectRoot, appResourcesPath);
+
+    const entryModule = nsWebpack.getEntryModule(appFullPath, platform);
+    const entryPath = `.${sep}${entryModule}`;
+    const entries = env.entries || {};
+    entries.bundle = entryPath;
+
+    const areCoreModulesExternal = Array.isArray(env.externals) && env.externals.some(e => e.indexOf('tns-core-modules') > -1);
+    if (platform === 'ios' && !areCoreModulesExternal) {
+        entries['tns_modules/tns-core-modules/inspector_modules'] = 'inspector_modules';
+    }
+    console.log(`Bundling application for entryPath ${entryPath}...`);
+
+    const sourceMapFilename = nsWebpack.getSourceMapFilename(hiddenSourceMap, __dirname, dist);
+
+    const itemsToClean = [`${dist}/**/*`];
+    if (platform === 'android') {
+        itemsToClean.push(`${join(projectRoot, 'platforms', 'android', 'app', 'src', 'main', 'assets', 'snapshots')}`);
+        itemsToClean.push(`${join(projectRoot, 'platforms', 'android', 'app', 'build', 'configurations', 'nativescript-android-snapshot')}`);
+    }
+
+    const package = require('./package.json');
+    const isIOS = platform === 'ios';
+    const isAndroid = platform === 'android';
+    const APP_STORE_ID = process.env.IOS_APP_ID;
+    const defines = mergeOptions(
+        {
+            PRODUCTION: !!production,
+            process: 'global.process',
+            'global.TNS_WEBPACK': 'true',
+            'gVars.platform': `"${platform}"`,
+            'gVars.isIOS': platform === 'ios',
+            'gVars.isAndroid': platform === 'android',
+            TNS_ENV: JSON.stringify(mode),
+            'gVars.sentry': !!sentry,
+            SENTRY_DSN: `"${process.env.SENTRY_DSN}"`,
+            SENTRY_PREFIX: `"${!!sentry ? process.env.SENTRY_PREFIX : ''}"`,
+            LOG_LEVEL: devlog ? '"full"' : '""',
+            TEST_LOGS: adhoc || !production,
+            GIT_URL: `"${package.repository}"`,
+            STORE_LINK: `"${
+                isAndroid
+                    ? `https://play.google.com/store/apps/details?id=${package.nativescript.id}`
+                    : `https://itunes.apple.com/app/id${APP_STORE_ID}`
+            }"`,
+            STORE_REVIEW_LINK: `"${
+                isIOS
+                    ? ` itms-apps://itunes.apple.com/WebObjects/MZStore.woa/wa/viewContentsUserReviews?id=${APP_STORE_ID}&onlyLatestVersion=true&pageNumber=0&sortOrdering=1&type=Purple+Software`
+                    : `market://details?id=${package.nativescript.id}`
+            }"`,
         },
+        params.definePlugin || {}
+    );
+    
+    console.log('defines', defines);
+
+    const symbolsParser = require('scss-symbols-parser');
+    const mdiSymbols = symbolsParser.parseSymbols(readFileSync(resolve(projectRoot, 'node_modules/@mdi/font/scss/_variables.scss')).toString());
+    const mdiIcons = JSON.parse(`{${mdiSymbols.variables[mdiSymbols.variables.length - 1].value.replace(/" (F|0)(.*?)([,\n]|$)/g, '": "$1$2"$3')}}`);
+
+    const scssPrepend = `$nunito-fontFamily: ${platform === 'android' ? 'res/nunito' : 'Nunito'};
+    $mdi-fontFamily: ${platform === 'android' ? 'materialdesignicons-webfont' : 'Material Design Icons'};`;
+
+    params.copyPlugin = [
+        { from: '../node_modules/@mdi/font/fonts/materialdesignicons-webfont.ttf', to: 'fonts' }
+    ];
+    nsWebpack.processAppComponents(appComponents, platform);
+    const config = {
+        mode,
+        context: appFullPath,
+        externals: externals.concat(params.externals || []),
+        watchOptions: {
+            ignored: [
+                appResourcesFullPath,
+                // Don't watch hidden files
+                '**/.*'
+            ]
+        },
+        target: nativescriptTarget,
+        // target: nativeScriptVueTarget,
+        entry: entries,
         output: {
             pathinfo: false,
             path: dist,
-            libraryTarget: "commonjs2",
-            filename: "[name].js",
-            globalObject: "global"
+            libraryTarget: 'commonjs2',
+            filename: '[name].js',
+            globalObject: 'global',
+            hashSalt
         },
         resolve: {
-            extensions: [".vue", ".ts", ".tsx", ".js", ".scss", ".css"],
+            extensions: ['.vue', '.mjs', '.ts', '.js', '.scss', '.css'],
             // Resolve {N} system modules from tns-core-modules
-            modules: [
-                resolve(__dirname, "node_modules/tns-core-modules"),
-                resolve(__dirname, "node_modules"),
-                "node_modules/tns-core-modules",
-                "node_modules"
-            ],
-            alias: {
-                "~": appFullPath,
-                vue: "nativescript-vue"
-            },
+            modules: [resolve(__dirname, `node_modules/${coreModulesPackageName}`), resolve(__dirname, 'node_modules'), `node_modules/${coreModulesPackageName}`, 'node_modules'],
+            alias,
             // resolve symlinks to symlinked modules
-            symlinks: true,
+            symlinks: true
         },
         resolveLoader: {
             // don't resolve symlinks to symlinked loaders
@@ -139,114 +250,198 @@ module.exports = env => {
             http: false,
             timers: false,
             setImmediate: false,
-            fs: "empty",
+            fs: 'empty',
+            crypto: 'empty',
             __dirname: false
         },
-        devtool: sourceMap ? "inline-source-map" : "none",
+        devtool: inlineSourceMap ? 'inline-cheap-source-map' : false,
         optimization: {
+            runtimeChunk: 'single',
+            noEmitOnErrors: true,
             splitChunks: {
                 cacheGroups: {
                     vendor: {
-                        name: "vendor",
-                        chunks: "all",
+                        name: 'vendor',
+                        chunks: 'all',
                         test: module => {
-                            const moduleName = module.nameForCondition
-                                ? module.nameForCondition()
-                                : ""
+                            const moduleName = module.nameForCondition ? module.nameForCondition() : '';
                             return (
                                 /[\\/]node_modules[\\/]/.test(moduleName) ||
-                                appComponents.some(comp => comp === moduleName)
-                            )
+                                /@nativescript\/core/.test(moduleName) ||
+                                /nativescript-core/.test(moduleName) || // this one is for linked nativescript core build
+                                appComponents.some(comp => comp === moduleName) ||
+                                (params.chunkTestCallback && params.chunkTestCallback(moduleName))
+                            );
                         },
                         enforce: true
                     }
                 }
             },
-            minimize: !!uglify,
+            minimize: uglify !== undefined ? uglify : production,
             minimizer: [
-                new TerserPlugin({
-                    parallel: true,
-                    cache: true,
-                    terserOptions: {
-                        output: {
-                            comments: false
+                new TerserPlugin(
+                    mergeOptions(
+                        {
+                            parallel: true,
+                            cache: true,
+                            sourceMap: isAnySourceMapEnabled,
+                            terserOptions: {
+                                ecma: 6,
+                                // warnings: true,
+                                // toplevel: true,
+                                output: {
+                                    comments: false,
+                                    semicolons: !isAnySourceMapEnabled
+                                },
+                                compress: {
+                                    // The Android SBG has problems parsing the output
+                                    // when these options are enabled
+                                    collapse_vars: platform !== 'android',
+                                    sequences: platform !== 'android',
+                                    passes: 2
+                                },
+                                keep_fnames: true
+                            }
                         },
-                        compress: {
-                            // The Android SBG has problems parsing the output
-                            // when these options are enabled
-                            collapse_vars: platform !== "android",
-                            sequences: platform !== "android"
-                        },
-                        safari10: platform === "ios"
-                    }
-                })
+                        params.terserOptions || {}
+                    )
+                )
             ]
         },
         module: {
             rules: [
                 {
-                    test: new RegExp(entryPath),
+                    include: [join(appFullPath, entryPath + '.js'), join(appFullPath, entryPath + '.ts')],
                     use: [
                         // Require all Android app components
-                        platform === "android" && {
-                            loader:
-                                "nativescript-dev-webpack/android-app-components-loader",
+                        platform === 'android' && {
+                            loader: resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/android-app-components-loader'),
                             options: { modules: appComponents }
                         },
 
                         {
-                            loader:
-                                "nativescript-dev-webpack/bundle-config-loader",
+                            loader: resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/bundle-config-loader'),
                             options: {
                                 registerPages: true, // applicable only for non-angular apps
-                                loadCss: !snapshot // load the application css if in debug mode
+                                loadCss: !snapshot, // load the application css if in debug mode
+                                unitTesting,
+                                appFullPath,
+                                projectRoot,
+                                ignoredFiles: nsWebpack.getUserDefinedEntries(entries, platform)
                             }
                         }
                     ].filter(loader => Boolean(loader))
                 },
                 {
-                    test: /\.css$/,
+                    // needs to be boefore the string replace
+                    test: /\.vue$/,
                     use: [
-                        "nativescript-dev-webpack/style-hot-loader",
-                        "nativescript-dev-webpack/apply-css-loader.js",
                         {
-                            loader: "css-loader",
-                            options: { url: false }
+                            loader: resolve(__dirname, 'node_modules', 'vue-loader'),
+                            options: {
+                                compiler: NsVueTemplateCompiler
+                            }
                         }
                     ]
                 },
                 {
-                    test: /\.scss$/,
+                    // rules to replace mdi icons and not use nativescript-font-icon
+                    test: /\.(ts|js|css|vue)$/,
+                    exclude: /node_modules/,
                     use: [
-                        "nativescript-dev-webpack/style-hot-loader",
-                        "nativescript-dev-webpack/apply-css-loader.js",
                         {
-                            loader: "css-loader",
-                            options: { url: false }
-                        },
-                        "sass-loader"
+                            loader: resolve(__dirname, 'node_modules', 'string-replace-loader'),
+                            options: {
+                                search: 'mdi-([a-z0-9-]+)',
+                                replace: (match, p1, offset) => {
+                                    if (mdiIcons[p1]) {
+                                        return String.fromCharCode(parseInt(mdiIcons[p1], 16));
+                                    }
+                                    return match;
+                                },
+                                flags: 'g'
+                            }
+                        }
                     ]
                 },
                 {
-                    test: /\.tsx?$/,
+                    test: /[\/|\\]app\.css$/,
+                    use: [
+                        resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/style-hot-loader'),
+                        {
+                            loader: resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/css2json-loader'),
+                            options: { useForImports: true }
+                        }
+                    ]
+                },
+                {
+                    test: /[\/|\\]app\.scss$/,
+                    use: [
+                        resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/style-hot-loader'),
+                        {
+                            loader: resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/css2json-loader'),
+                            options: { useForImports: true }
+                        },
+                        {
+                            loader: resolve(__dirname, 'node_modules', 'sass-loader'),
+                            options: {
+                                sourceMap: false,
+                                prependData: scssPrepend
+                            }
+                        }
+                    ]
+                },
+                {
+                    test: /\.css$/,
+                    exclude: /[\/|\\]app\.css$/,
+                    use: [
+                        resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/style-hot-loader'),
+                        resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/apply-css-loader.js'),
+                        { loader: resolve(__dirname, 'node_modules', 'css-loader'), options: { url: false } }
+                    ]
+                },
+                {
+                    test: /\.scss$/,
+                    exclude: /[\/|\\]app\.scss$/,
+                    use: [
+                        resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/style-hot-loader'),
+                        resolve(__dirname, 'node_modules', 'nativescript-dev-webpack/apply-css-loader.js'),
+                        { loader: resolve(__dirname, 'node_modules', 'css-loader'), options: { url: false } },
+                        {
+                            loader: resolve(__dirname, 'node_modules', 'sass-loader'),
+                            options: {
+                                sourceMap: false,
+                                prependData: scssPrepend
+                            }
+                        }
+                    ]
+                },
+                {
+                    test: /\.mjs$/,
+                    type: 'javascript/auto'
+                },
+                // {
+                //     test: /\.js$/,
+                //     loader: 'babel-loader'
+                // },
+                {
+                    test: /\.ts$/,
                     exclude: /node_modules/,
-                    loader: "ts-loader",
-                    options: {
-                        configFile: resolve(tsconfig),
-                        appendTsSuffixTo: [/\.vue$/],
-                        allowTsInNodeModules: true // wanted?
-                    }
-                },
-                {
-                    test: /\.js$/,
-                    loader: "babel-loader"
-                },
-                {
-                    test: /\.vue$/,
-                    loader: "vue-loader",
-                    options: {
-                        compiler: NsVueTemplateCompiler
-                    }
+                    use: [
+                        {
+                            loader: resolve(__dirname, 'node_modules', 'ts-loader'),
+                            options: {
+                                configFile: resolve(tsconfig),
+                                transpileOnly: true,
+                                appendTsSuffixTo: [/\.vue$/],
+                                allowTsInNodeModules: true,
+                                compilerOptions: {
+                                    sourceMap: isAnySourceMapEnabled,
+                                    declaration: false
+                                }
+                            }
+                        }
+                    ]
                 }
             ]
         },
@@ -255,32 +450,37 @@ module.exports = env => {
             // make sure to include the plugin!
             new VueLoaderPlugin(),
             // Define useful constants like TNS_WEBPACK
-            new webpack.DefinePlugin({
-                process: undefined,
-                "global.TNS_WEBPACK": "true",
-                TNS_ENV: JSON.stringify(mode)
-            }),
+            new webpack.EnvironmentPlugin(
+                mergeOptions(
+                    {
+                        NODE_ENV: JSON.stringify(mode), // use 'development' unless process.env.NODE_ENV is defined
+                        DEBUG: false
+                    },
+                    params.environmentPlugin || {}
+                )
+            ),
+            new webpack.DefinePlugin(defines),
             // Remove all files from the out dir.
-            new CleanWebpackPlugin([`${dist}/**/*`]),
-            // Copy native app resources to out dir.
-            new CopyWebpackPlugin([
-                {
-                    from: `${appResourcesFullPath}/${appResourcesPlatformDir}`,
-                    to: `${dist}/App_Resources/${appResourcesPlatformDir}`,
-                    context: projectRoot
-                }
-            ]),
+            new CleanWebpackPlugin({
+                dangerouslyAllowCleanPatternsOutsideProject: true,
+                dry: false,
+                verbose: !!verbose,
+                cleanOnceBeforeBuildPatterns: itemsToClean
+            }),
             // Copy assets to out dir. Add your own globs as needed.
             new CopyWebpackPlugin(
                 [
-                    { from: "fonts/**" },
-                    { from: "**/*.+(jpg|png)" },
-                    { from: "assets/**/*" }
-                ],
-                { ignore: [`${relative(appPath, appResourcesFullPath)}/**`] }
+                    { from: 'fonts/!(ios|android)/**/*', to: 'fonts', flatten: true },
+                    { from: 'fonts/*', to: 'fonts', flatten: true },
+                    { from: `fonts/${platform}/**/*`, to: 'fonts', flatten: true },
+                    { from: '**/*.+(jpg|png)' },
+                    { from: 'assets/**/*' }
+                ].concat(params.copyPlugin || []),
+                {
+                    ignore: [`${relative(appPath, appResourcesFullPath)}/**`]
+                }
             ),
-            // Generate a bundle starter script and activate it in package.json
-            new nsWebpack.GenerateBundleStarterPlugin(["./vendor", "./bundle"]),
+            new nsWebpack.GenerateNativeScriptEntryPointsPlugin('bundle'),
             // For instructions on how to set up workers with webpack
             // check out https://github.com/nativescript/worker-loader
             new NativeScriptWorkerPlugin(),
@@ -291,36 +491,130 @@ module.exports = env => {
             // Does IPC communication with the {N} CLI to notify events when running in watch mode.
             new nsWebpack.WatchStateLoggerPlugin()
         ]
+    };
+    if (hiddenSourceMap || sourceMap) {
+        if (!!sentry) {
+            // const sourceMapFilename = nsWebpack.getSourceMapFilename(hiddenSourceMap, __dirname, dist);
+
+            config.plugins.push(
+                new webpack.SourceMapDevToolPlugin(
+                    mergeOptions(
+                        {
+                            append: `\n//# sourceMappingURL=${process.env.SENTRY_PREFIX}[name].js.map`,
+                            filename: join(process.env.SOURCEMAP_REL_DIR, '[name].js.map')
+                        },
+                        params.sourceMapPlugin || {}
+                    )
+                )
+            );
+            let appVersion;
+            let buildNumber;
+            if (platform === 'android') {
+                appVersion = readFileSync('app/App_Resources/Android/app.gradle', 'utf8').match(/versionName "((?:[0-9]+\.?)+)"/)[1];
+                buildNumber = readFileSync('app/App_Resources/Android/app.gradle', 'utf8').match(/versionCode ([0-9]+)/)[1];
+            } else if (platform === 'ios') {
+                appVersion = readFileSync('app/App_Resources/iOS/Info.plist', 'utf8').match(/<key>CFBundleShortVersionString<\/key>[\s\n]*<string>(.*?)<\/string>/)[1];
+                buildNumber = readFileSync('app/App_Resources/iOS/Info.plist', 'utf8').match(/<key>CFBundleVersion<\/key>[\s\n]*<string>([0-9]*)<\/string>/)[1];
+            }
+            console.log('appVersion', appVersion, buildNumber);
+
+            config.plugins.push(
+                new SentryCliPlugin({
+                    release: appVersion,
+                    urlPrefix: 'app:///',
+                    rewrite: true,
+                    dist: `${buildNumber}.${platform}`,
+                    ignore: ['tns-java-classes', 'hot-update'],
+                    include: [dist, join(dist, process.env.SOURCEMAP_REL_DIR)]
+                })
+            );
+        } else {
+            config.plugins.push(
+                new webpack.SourceMapDevToolPlugin(
+                    mergeOptions(
+                        {
+                            noSources: true
+                            //  fileContext:params.sourceMapPublicPath
+                            // publicPath: params.sourceMapPublicPath || '~/',
+                            // fileContext:sourceMapsDist,
+                            // noSources:true,
+                            // moduleFilenameTemplate: info => {
+                            //     let filePath = info.identifier;
+                            //     console.log( 'filePath', filePath);
+                            //     if (filePath.startsWith('./')) {
+                            //         filePath = './app' + filePath.slice(1);
+                            //     }
+                            //     // if (filePath.startsWith('file:///')) {
+                            //     //     filePath = './app' + filePath.slice(1);
+                            //     // }
+                            //     return filePath.replace('file:///', '');
+                            // }
+                        },
+                        params.sourceMapPlugin || {}
+                    )
+                )
+            );
+        }
+    }
+
+    if (unitTesting) {
+        config.module.rules.push(
+            {
+                test: /-page\.js$/,
+                use: 'nativescript-dev-webpack/script-hot-loader'
+            },
+            {
+                test: /\.(html|xml)$/,
+                use: 'nativescript-dev-webpack/markup-hot-loader'
+            },
+
+            { test: /\.(html|xml)$/, use: 'nativescript-dev-webpack/xml-namespace-loader' }
+        );
     }
 
     if (!!report) {
         // Generate report files for bundles content
         config.plugins.push(
             new BundleAnalyzerPlugin({
-                analyzerMode: "static",
+                analyzerMode: 'static',
                 openAnalyzer: false,
-                generateStatsFile: true,
-                reportFilename: resolve(projectRoot, "report", `report.html`),
-                statsFilename: resolve(projectRoot, "report", `stats.json`)
+                reportFilename: resolve(projectRoot, 'report', 'report.html')
             })
-        )
+        );
     }
 
     if (!!snapshot) {
-        config.plugins.push(
-            new nsWebpack.NativeScriptSnapshotPlugin({
-                chunk: "vendor",
-                requireModules: ["tns-core-modules/bundle-entry-points"],
+        const options = mergeOptions(
+            {
+                chunk: 'vendor',
+                requireModules: ['tns-core-modules/bundle-entry-points'],
                 projectRoot,
-                targetArchs: ['arm'],
-                webpackConfig: config
-            })
-        )
+                // targetArchs: params.targetArchs || ['arm'],
+                snapshotInDocker,
+                skipSnapshotTools,
+                useLibs
+            },
+            params.snapshotPlugin
+        );
+        config.plugins.push(new nsWebpack.NativeScriptSnapshotPlugin(Object.assign(options, { webpackConfig: config })));
     }
 
     if (!!hmr) {
-        config.plugins.push(new webpack.HotModuleReplacementPlugin())
+        config.plugins.push(new webpack.HotModuleReplacementPlugin());
     }
 
-    return config
-}
+    // if (!!production) {
+    //     config.plugins.push(
+    //         new ForkTsCheckerWebpackPlugin({
+    //             tsconfig: resolve(tsconfig),
+    //             async: false,
+    //             useTypescriptIncrementalApi: true,
+    //             checkSyntacticErrors: true,
+    //             memoryLimit: 4096
+    //             // workers: 1
+    //         })
+    //     );
+    // }
+
+    return config;
+};
